+++ jdaviz/configs/imviz/plugins/viewers.py
@@ -25,14 +25,15 @@ class ImvizImageView(BqplotImageView, AstrowidgetsImageViewerMixin, JdavizViewer
     tools = ['jdaviz:homezoom', 'jdaviz:boxzoommatch', 'jdaviz:boxzoom',
              'jdaviz:panzoommatch', 'jdaviz:imagepanzoom',
              'jdaviz:contrastbias', 'jdaviz:blinkonce',
-             'bqplot:rectangle', 'bqplot:circle', 'bqplot:ellipse']
+             'bqplot:rectangle', 'bqplot:circle', 'bqplot:ellipse', 'jdaviz:singlepixelregion']
 
     # categories: zoom resets, zoom, pan, subset, select tools, shortcuts
     tools_nested = [
                     ['jdaviz:homezoom', 'jdaviz:prevzoom'],
                     ['jdaviz:boxzoommatch', 'jdaviz:boxzoom'],
                     ['jdaviz:panzoommatch', 'jdaviz:imagepanzoom'],
-                    ['bqplot:circle', 'bqplot:rectangle', 'bqplot:ellipse'],
+                    ['bqplot:circle', 'bqplot:rectangle', 'bqplot:ellipse',
+                     'jdaviz:singlepixelregion'],
                     ['jdaviz:blinkonce', 'jdaviz:contrastbias'],
                     ['jdaviz:sidebar_plot', 'jdaviz:sidebar_export', 'jdaviz:sidebar_compass']
                 ]
@@ -50,8 +51,7 @@ class ImvizImageView(BqplotImageView, AstrowidgetsImageViewerMixin, JdavizViewer
         self.line_profile_xy = None
 
         self.add_event_callback(self.on_mouse_or_key_event, events=['mousemove', 'mouseenter',
-                                                                    'mouseleave', 'keydown',
-                                                                    'click'])
+                                                                    'mouseleave', 'keydown'])
         self.state.add_callback('x_min', self.on_limits_change)
         self.state.add_callback('x_max', self.on_limits_change)
         self.state.add_callback('y_min', self.on_limits_change)
@@ -171,22 +171,6 @@ class ImvizImageView(BqplotImageView, AstrowidgetsImageViewerMixin, JdavizViewer
                 self.line_profile_xy.selected_y = y
                 self.line_profile_xy.selected_viewer = self.reference_id
                 self.line_profile_xy.vue_draw_plot()
-
-        elif data['event'] == 'click' and data['altKey'] is True:
-            from regions import RectanglePixelRegion, PixCoord
-
-            image = active_layer.layer
-
-            # Extract data coordinates - these are pixels in the reference image
-            x = data['domain']['x']
-            y = data['domain']['y']
-
-            # NOTE: We always use the reference image pixel coordinates because
-            # Subset is defined w.r.t. reference image.
-            x, y = np.rint([x, y])
-
-            reg = RectanglePixelRegion(center=PixCoord(x=x, y=y), width=1, height=1)
-            self.jdaviz_helper.load_regions(reg)
 
     def blink_once(self, reversed=False):
         # Simple blinking of images - this will make it so that only one