+++ jdaviz/configs/imviz/tests/test_tools.py
@@ -1,4 +1,5 @@
 import numpy as np
+from regions import RectanglePixelRegion
 
 from jdaviz.configs.imviz.tests.utils import BaseImviz_WCS_WCS
 
@@ -60,18 +61,45 @@ class TestPanZoomTools(BaseImviz_WCS_WCS):
         t_linkedpan.deactivate()
 
 
+# We use a new test class to avoid a dirty state from previous test.
+class TestSinglePixelRegion(BaseImviz_WCS_WCS):
+    def test_singlepixelregion(self):
+        self.imviz.link_data(link_type='wcs')
+
+        t = self.imviz.default_viewer.toolbar_nested.tools['jdaviz:singlepixelregion']
+        t.activate()
+
+        # Create a region while viewing reference data.
+        t.on_mouse_event({'event': 'click', 'domain': {'x': 1, 'y': 2}})
+        regions = self.imviz.get_interactive_regions()
+        assert len(regions) == 1
+        reg = regions['Subset 1']
+        assert (isinstance(reg, RectanglePixelRegion) and reg.center.x == 1 and reg.center.y == 2
+                and reg.width == 1 and reg.height == 1)
+
+        # Create a region while viewing dithered data.
+        # Region will still be w.r.t. reference data, that is, x and y from domain stay the same.
+        self.imviz.default_viewer.blink_once()
+        t.on_mouse_event({'event': 'click', 'domain': {'x': 3, 'y': 4}})
+        regions = self.imviz.get_interactive_regions()
+        assert len(regions) == 2
+        reg = regions['Subset 2']
+        assert (isinstance(reg, RectanglePixelRegion) and reg.center.x == 3 and reg.center.y == 4
+                and reg.width == 1 and reg.height == 1)
+
+        t.deactivate()
+
+
 def test_blink(imviz_helper):
     viewer = imviz_helper.default_viewer
 
     for i in range(3):
         imviz_helper.load_data(np.zeros((2, 2)) + i, data_label=f'image_{i}')
 
-    # Last loaded is shown first. So, blinking will take you back to the first one.
-    # Blink forward. The event will also initialize viewer.label_mouseover .
     viewer.on_mouse_or_key_event({'event': 'keydown', 'key': 'b', 'domain': {'x': 0, 'y': 0}})
     assert viewer.label_mouseover.value == '+0.00000e+00 '
 
-    # Blink forward again and update coordinates info panel.
+    # Blink forward and update coordinates info panel.
     viewer.blink_once()
     viewer.on_mouse_or_key_event({'event': 'mousemove', 'domain': {'x': 0, 'y': 0}})
     assert viewer.label_mouseover.value == '+1.00000e+00 '